# Conflicts:
#	exonum/src/blockchain/block.rs

+++ exonum/src/runtime/dispatcher/mod.rs
@@ -317,7 +317,7 @@ impl Dispatcher {
         &self,
         fork: &mut Fork,
         tx_id: Hash,
-        tx_index: u64,
+        tx_index: u32,
         tx: &Verified<AnyTx>,
     ) -> Result<(), ExecutionError> {
         let caller = Caller::Transaction {
